+++ src/apps/graphics/libgpuperfcnt.mk
@@ -12,8 +12,10 @@ libgpuperfcnt:
 	 $(call fbprint_b,"libgpuperfcnt") && \
 	 cd $(GRAPHICSDIR) && \
 	 if [ ! -d libgpuperfcnt ]; then \
-	     wget -q $(repo_libgpuperfcnt_bin_url) -O libgpuperfcnt.bin $(LOG_MUTE) && \
-	     chmod +x libgpuperfcnt.bin && ./libgpuperfcnt.bin --auto-accept $(LOG_MUTE) && \
+		 rm -rf libgpuperfcnt*; \
+	     $(WGET) $(repo_libgpuperfcnt_bin_url) -O libgpuperfcnt.bin $(LOG_MUTE); \
+		 [ $$? -ne 0 ] && { echo "Downloading $(repo_libgpuperfcnt_bin_url) failed."; exit 1; } || \
+	     chmod +x libgpuperfcnt.bin && ./libgpuperfcnt.bin --auto-accept --force $(LOG_MUTE); \
 	     mv libgpuperfcnt-* libgpuperfcnt && rm -f libgpuperfcnt.bin; \
 	 fi && \
 	 cp -Prf libgpuperfcnt/usr $(DESTDIR) && \
